Rename folder enitity -> entity
