+++ src/main/java/com/github/legoatoom/connectiblechains/client/ClientInitializer.java
@@ -24,9 +24,9 @@ import com.github.legoatoom.connectiblechains.client.render.entity.ChainKnotEnti
 import com.github.legoatoom.connectiblechains.client.render.entity.ChainTextureManager;
 import com.github.legoatoom.connectiblechains.client.render.entity.model.ChainKnotEntityModel;
 import com.github.legoatoom.connectiblechains.config.ModConfig;
-import com.github.legoatoom.connectiblechains.enitity.ChainCollisionEntity;
-import com.github.legoatoom.connectiblechains.enitity.ChainKnotEntity;
-import com.github.legoatoom.connectiblechains.enitity.ModEntityTypes;
+import com.github.legoatoom.connectiblechains.entity.ChainCollisionEntity;
+import com.github.legoatoom.connectiblechains.entity.ChainKnotEntity;
+import com.github.legoatoom.connectiblechains.entity.ModEntityTypes;
 import com.github.legoatoom.connectiblechains.util.Helper;
 import com.github.legoatoom.connectiblechains.util.NetworkingPackets;
 import me.shedaniel.autoconfig.AutoConfig;
